+++ src/infra/db/prisma/external-auth-mapping/index.ts
@@ -1,0 +1,2 @@
+export * from './add-external-auth-mapping/add-external-auth-mapping-prisma-repo'
+export * from './find-external-auth-mapping/find-external-auth-mapping-by-external-auth-user-id-prisma-repo'
