+++ src/main/factories/infra/db/prisma/external-auth-mapping/add-external-auth-mapping-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { AddExternalAuthMappingPrismaRepo } from '@/infra/db/prisma/external-auth-mapping/add-external-auth-mapping/add-external-auth-mapping-prisma-repo'
+import { AddExternalAuthMappingPrismaRepo } from '@/infra/db/prisma/external-auth-mapping'
 import { type AddExternalAuthMappingRepo } from '@/usecases/contracts/db/external-auth-mapping'
 
 export const makeAddExternalAuthMappingPrismaRepo = (): AddExternalAuthMappingRepo => {
