+++ src/main/factories/infra/db/prisma/external-auth-mapping/find-external-auth-mapping-by-external-auth-user-id-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { FindExternalAuthMappingByExternalAuthUserIdPrismaRepo } from '@/infra/db/prisma/external-auth-mapping/find-external-auth-mapping/find-external-auth-mapping-by-external-auth-user-id-prisma-repo'
+import { FindExternalAuthMappingByExternalAuthUserIdPrismaRepo } from '@/infra/db/prisma/external-auth-mapping'
 import { type FindExternalAuthMappingByExternalAuthUserIdRepo } from '@/usecases/contracts/db/external-auth-mapping'
 
 export const makeFindExternalAuthMappingByExternalAuthUserIdPrismaRepo = (): FindExternalAuthMappingByExternalAuthUserIdRepo => {
