+++ src/main/factories/middlewares/auth-middleware-factory.ts
@@ -1,5 +1,5 @@
 import { type Middleware } from '@/presentation/contracts'
-import { AuthMiddleware } from '@/presentation/middlewares/auth/auth-middleware'
+import { AuthMiddleware } from '@/presentation/middlewares'
 import { makeAuthUseCase } from '../usecases/auth/auth-usecase-factory'
 
 export const makeAuthMiddleware = (): Middleware => {
