+++ src/main/factories/middlewares/webhook-validator-middleware-factory.ts
@@ -1,5 +1,5 @@
 import { type Middleware } from '@/presentation/contracts'
-import { WebhookValidatorMiddleware } from '@/presentation/middlewares/web-hook/webhook-validator-middleware'
+import { WebhookValidatorMiddleware } from '@/presentation/middlewares'
 import { SvixWebhookValidation } from '@/validators/webhook/svix-webhook-validation'
 
 export const makeWebhookValidatorMiddleware = (): Middleware => {
