+++ src/main/routes/user/signup/signup-route.ts
@@ -1,5 +1,5 @@
 import { adaptRoute } from '@/main/factories/adapters/nest-route-adapter-factory'
-import { makeSignUpController } from '@/main/factories/controllers/user/signup/signup-controller-factory'
+import { makeSignUpController } from '@/main/factories/controllers/user/signup-controller-factory'
 import { makeAdaptClerkRequestSignUpControllerDecorator } from '@/main/factories/decorators'
 import { Controller, Post, Req, Res } from '@nestjs/common'
 import { Request, Response } from 'express'
